+++ src/main/java/edu/agh/streamgraph/dynsssp/EdgeWrapper.java
@@ -5,11 +5,11 @@ import org.apache.flink.graph.Edge;
 import org.apache.flink.types.NullValue;
 import org.apache.flink.util.Collector;
 
-public class EdgeWrapper implements FlatMapFunction<Edge<Long, NullValue> , ProcessMessage> {
+public class EdgeWrapper implements FlatMapFunction<Edge<Integer, NullValue> , ProcessMessage> {
 
     @Override
-    public void flatMap(Edge<Long, NullValue> edge, Collector<ProcessMessage> collector) throws Exception {
+    public void flatMap(Edge<Integer, NullValue> edge, Collector<ProcessMessage> collector) throws Exception {
         collector.collect(ProcessMessage.forVertexAddition(edge));
-        collector.collect(ProcessMessage.forVertexAddition(new Edge<Long, NullValue> (edge.f1, edge.f0, NullValue.getInstance())));
+        collector.collect(ProcessMessage.forVertexAddition(new Edge<Integer, NullValue> (edge.f1, edge.f0, NullValue.getInstance())));
     }
 }
